files update

+++ Uber/contracts/src/main/java/com/team21/uber/contracts/dto/RideSummaryDTO.java
@@ -1,13 +1,13 @@
 package com.team21.uber.contracts.dto;
 
 public record RideSummaryDTO(
-        long totalRides,
-        long completedRides,
-        long cancelledRides,
+        int totalRides,
+        int completedRides,
+        int cancelledRides,
         double totalSpent,
         double averageFare
 ) {
     public static RideSummaryDTO empty() {
-        return new RideSummaryDTO(0L, 0L, 0L, 0.0, 0.0);
+        return new RideSummaryDTO(0, 0, 0, 0.0, 0.0);
     }
 }
